move OrderStatus, MarketInfo, etc. to market/types package
